+++ libs/journeys/ui/src/components/TemplateGalleryCard/TemplateGalleryCard.spec.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event'
 import { NextRouter, useRouter } from 'next/router'
 import { TemplateGalleryCard } from '.'
 import '../../../test/i18n'
-import { algoliaJourneys } from '../../libs/algolia/useAlgoliaJourneys/data'
+import { algoliaJourneys } from '../../libs/algolia/useAlgoliaJourneys/useAlgoliaJourneys.mock'
 
 jest.mock('next/router', () => ({
   __esModule: true,
